Integrate using IBM RXN 4 Chemistry Wrapper in API
Currently creates a new project every time. Only predicts product
based on reactants, but successfully provides a prediction of the
product with the results being only of:
1) message, 2) reactants, 3) predicted product, 4) confidence.

Currently implements polling to predict reaction until success.
Each run of the predict_product API call currently does 2 API
requests to the RXN4Chemistry wrapper, one for creating the project,
and one for predicting the product.